Minimally satisfactory refactoring.
It still adds state (an instance variable) to the User entity rather
than simply modifying behaviour, but then again, so does the
`ProfileFormatterSetup` immediately following that call, and I didn't
spin my gears all day writing *that*. (In my own defence, the latter
*is* understandably short and clear.)

RSpec: 713 examples, 0 failures, 1 pending
RuboCop: 230 files inspected, no offenses detected
